+++ src/main/kotlin/us/frollo/frollosdk/model/api/aggregation/providers/ProviderResponse.kt
@@ -17,6 +17,7 @@
 package us.frollo.frollosdk.model.api.aggregation.providers
 
 import com.google.gson.annotations.SerializedName
+import us.frollo.frollosdk.model.coredata.aggregation.providers.AggregatorType
 import us.frollo.frollosdk.model.coredata.aggregation.providers.ProviderAuthType
 import us.frollo.frollosdk.model.coredata.aggregation.providers.ProviderEncryption
 import us.frollo.frollosdk.model.coredata.aggregation.providers.ProviderLoginForm
@@ -32,9 +33,11 @@ internal data class ProviderResponse(
     @SerializedName("popular") val popular: Boolean,
     @SerializedName("container_names") val containerNames: List<String>,
     @SerializedName("login_url") val loginUrl: String?,
-
     @SerializedName("large_logo_url") val largeLogoUrl: String?,
     @SerializedName("large_logo_revision") val largeLogoRevision: Int?,
+    @SerializedName("aggregator_type") val aggregatorType: AggregatorType,
+    @SerializedName("permissions") val permissions: List<String>?,
+
     @SerializedName("base_url") val baseUrl: String?,
     @SerializedName("forget_password_url") val forgetPasswordUrl: String?,
     @SerializedName("o_auth_site") val oAuthSite: Boolean?,